+++ src/flows/CostCalculator/CostCalculatorDisclaimer.tsx
@@ -11,14 +11,15 @@ import {
 import { ScrollArea } from '@/src/components/ui/scroll-area';
 import { X } from 'lucide-react';
 import { Button } from '@/src/components/ui/button';
-import { useCostCalculatorDisclaimer } from '@/src/flows/CostCalculator/hooks';
+import { disclaimerData } from '@/src/flows/CostCalculator/disclaimerUtils';
 
-type DisclaimerProps = {
+type CostCalculatorDisclaimerProps = {
   label?: string;
 };
 
-export const Disclaimer = ({ label = 'Disclaimer' }: DisclaimerProps) => {
-  const { data: disclaimer } = useCostCalculatorDisclaimer();
+export const CostCalculatorDisclaimer = ({
+  label = 'Disclaimer',
+}: CostCalculatorDisclaimerProps) => {
   return (
     <Drawer>
       <DrawerTrigger asChild>
@@ -32,12 +33,12 @@ export const Disclaimer = ({ label = 'Disclaimer' }: DisclaimerProps) => {
             <X className="h-4 w-4" />
             <span className="sr-only">Close</span>
           </DrawerClose>
-          <DrawerTitle>{disclaimer?.data.title}</DrawerTitle>
+          <DrawerTitle>{disclaimerData?.data.title}</DrawerTitle>
           <DrawerDescription>
             For more details read our{' '}
             <Button variant="link" size="link" asChild>
               <a
-                href={disclaimer?.data.html_url}
+                href={disclaimerData?.data.html_url}
                 target="_blank"
                 rel="noopener noreferrer"
               >
@@ -49,7 +50,9 @@ export const Disclaimer = ({ label = 'Disclaimer' }: DisclaimerProps) => {
         <ScrollArea className="px-4 pb-4 overflow-y-auto max-h-[calc(80vh-120px)] cost-calculator-disclaimer-drawer-scroll-area">
           <div
             className="cost-calculator-disclaimer-drawer-body"
-            dangerouslySetInnerHTML={{ __html: disclaimer?.data.body ?? '' }}
+            dangerouslySetInnerHTML={{
+              __html: disclaimerData?.data.body ?? '',
+            }}
           ></div>
         </ScrollArea>
       </DrawerContent>
