+++ src/flows/CostCalculator/index.ts
@@ -3,12 +3,8 @@ export { CostCalculatorFlow } from './CostCalculatorFlow';
 export { CostCalculatorForm } from './CostCalculatorForm';
 export { CostCalculatorSubmitButton } from './CostCalculatorSubmitButton';
 export { CostCalculatorResetButton } from './CostCalculatorResetButton';
-export { Disclaimer as CostCalculatorDisclaimer } from './Disclaimer/Disclaimer';
-export {
-  useCostCalculator,
-  useCostCalculatorDisclaimer,
-  useCostCalculatorEstimationPdf,
-} from './hooks';
-
+export { CostCalculatorDisclaimer } from './CostCalculatorDisclaimer';
+export { useCostCalculator, useCostCalculatorEstimationPdf } from './hooks';
+export { disclaimerData } from './disclaimerUtils';
 export { CostCalculatorResults } from './Results/CostCalculatorResults';
 export { buildPayload as buildCostCalculatorEstimationPayload } from './utils';
